// model/checkoutCustomerFields.ts
/**
 * Rvvup API
 * Rvvup Public API
 *
 * The version of the OpenAPI document: 2024-03-01
 * Contact: info@rvvup.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';
import { CheckoutCustomerFieldType } from './checkoutCustomerFieldType';

export class CheckoutCustomerFields {
    /**
    * The optional customer fields for the checkout.
    */
    'optional'?: Set<CheckoutCustomerFieldType>;
    /**
    * The required customer fields for the checkout.
    */
    'required'?: Set<CheckoutCustomerFieldType>;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "optional",
            "baseName": "optional",
            "type": "Set<CheckoutCustomerFieldType>"
        },
        {
            "name": "required",
            "baseName": "required",
            "type": "Set<CheckoutCustomerFieldType>"
        }    ];

    static getAttributeTypeMap() {
        return CheckoutCustomerFields.attributeTypeMap;
    }
}